+++ src/wrappers/themis/python/pythemis/__init__.py
@@ -15,8 +15,54 @@
 # limitations under the License.
 #
 
+import os
+import sys
+import warnings
 from ctypes import cdll
 from ctypes.util import find_library
+
+
+def _canonical_themis_paths():
+    paths = []
+
+    if sys.platform.startswith('linux'):
+        # Default library installation path for "make install"
+        paths.append('/usr/local/lib/libthemis.so.0')
+        # Don't bother figuring out the "right" absolute path, since
+        # that depends on the distro and non-distro-provided Pythons
+        # don't know about the right paths at all. Simply use soname
+        # and let ld figure it out.
+        paths.append('libthemis.so.0')
+
+    if sys.platform.startswith('darwin'):
+        # Default library installation path for "make install"
+        paths.append('/usr/local/lib/libthemis.0.dylib')
+        # These are install names of libraries installed via Homebrew
+        # Add both M1 and Intel paths so that x86 Ruby works on M1
+        paths.append('/opt/homebrew/opt/libthemis/lib/libthemis.0.dylib')
+        paths.append('/usr/local/opt/libthemis/lib/libthemis.0.dylib')
+        # Last try, look for ABI-qualified name
+        paths.append('libthemis.0.dylib')
+
+    return paths
+
+
+def _load_themis():
+    for path in _canonical_themis_paths():
+        try:
+            return cdll.LoadLibrary(path)
+        except: # TODO
+            continue
+
+    warnings.warn("""failed to load the canonical Themis Core library
+
+Proceeding to find 'themis' library in standard paths.
+This might cause ABI mismatch and crash the process.
+""",
+                  category=RuntimeWarning)
+
+    return cdll.LoadLibrary(find_library('themis'))
+
 
 __all__ = [
     "scell",
@@ -26,4 +72,4 @@ __all__ = [
     "ssession",
 ]
 
-themis = cdll.LoadLibrary(find_library('themis'))
+themis = _load_themis()